refactor: create service filter
