+++ src/app/core/models/form/form-product-filter.model.ts
@@ -1,0 +1,11 @@
+import { FormControl } from '@angular/forms';
+
+export class FormProductFilter {
+	constructor(
+		public category: string[],
+		public color: string[],
+		public tag: string[],
+		public price: number[],
+		public weight: string[],
+	) {}
+}
